CPLP-1227 refactor existing identityprovider user upload to new framework implementation
