+++ src/framework/Framework.IO/Framework.IO.csproj
@@ -22,6 +22,9 @@
   <ItemGroup>
     <ProjectReference Include="..\Framework.ErrorHandling\Framework.ErrorHandling.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="System.Linq.Async" Version="6.0.1" />
+  </ItemGroup>
   <PropertyGroup>
     <AssemblyName>Org.CatenaX.Ng.Portal.Backend.Framework.IO</AssemblyName>
     <TargetFramework>net6.0</TargetFramework>
